requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11
django-js-asset==1.2.2
django-mptt==0.9.1
idna==2.8
oauthlib==3.0.1
pkg-resources==0.0.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-openid==2.2.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
